moved project settings pages to a subpackage
