+++ src/main/java/ca/corefacility/bioinformatics/irida/ria/web/projects/settings/ProjectSettingsController.java
@@ -1,4 +1,4 @@
-package ca.corefacility.bioinformatics.irida.ria.web.projects;
+package ca.corefacility.bioinformatics.irida.ria.web.projects.settings;
 
 import ca.corefacility.bioinformatics.irida.exceptions.IridaWorkflowNotFoundException;
 import ca.corefacility.bioinformatics.irida.model.joins.impl.ProjectMetadataTemplateJoin;
@@ -11,6 +11,8 @@ import ca.corefacility.bioinformatics.irida.model.user.User;
 import ca.corefacility.bioinformatics.irida.model.workflow.IridaWorkflow;
 import ca.corefacility.bioinformatics.irida.model.workflow.analysis.type.AnalysisType;
 import ca.corefacility.bioinformatics.irida.model.workflow.submission.AnalysisSubmissionTemplate;
+import ca.corefacility.bioinformatics.irida.ria.web.projects.ProjectControllerUtils;
+import ca.corefacility.bioinformatics.irida.ria.web.projects.ProjectsController;
 import ca.corefacility.bioinformatics.irida.service.AnalysisSubmissionService;
 import ca.corefacility.bioinformatics.irida.service.ProjectService;
 import ca.corefacility.bioinformatics.irida.service.remote.ProjectRemoteService;
@@ -39,8 +41,6 @@ public class ProjectSettingsController {
 	private final MetadataTemplateService metadataTemplateService;
 	private final ProjectControllerUtils projectControllerUtils;
 	private final ProjectService projectService;
-	private final ProjectRemoteService projectRemoteService;
-	private final UserService userService;
 	private AnalysisSubmissionService analysisSubmissionService;
 	private IridaWorkflowsService workflowsService;
 
@@ -49,13 +49,11 @@ public class ProjectSettingsController {
 	@Autowired
 	public ProjectSettingsController(MessageSource messageSource, MetadataTemplateService metadataTemplateService,
 			ProjectControllerUtils projectControllerUtils, ProjectService projectService,
-			ProjectRemoteService projectRemoteService, UserService userService, AnalysisSubmissionService analysisSubmissionService, IridaWorkflowsService workflowsService) {
+			AnalysisSubmissionService analysisSubmissionService, IridaWorkflowsService workflowsService) {
 		this.messageSource = messageSource;
 		this.metadataTemplateService = metadataTemplateService;
 		this.projectControllerUtils = projectControllerUtils;
 		this.projectService = projectService;
-		this.projectRemoteService = projectRemoteService;
-		this.userService = userService;
 		this.analysisSubmissionService = analysisSubmissionService;
 		this.workflowsService = workflowsService;
 	}
@@ -177,30 +175,7 @@ public class ProjectSettingsController {
 		return "redirect:/projects/" + projectId + "/settings";
 	}
 
-	/**
-	 * Request for a {@link Project} remote settings page
-	 *
-	 * @param projectId
-	 *            the ID of the {@link Project} to read
-	 * @param model
-	 *            Model for the view
-	 * @param principal
-	 *            Logged in user
-	 *
-	 * @return name of the project remote settings page
-	 */
-	@RequestMapping("/remote")
-	@PreAuthorize("hasPermission(#projectId, 'canManageLocalProjectSettings')")
-	public String getProjectSettingsRemotePage(@PathVariable Long projectId, final Model model,
-			final Principal principal) {
-		Project project = projectService.read(projectId);
-		model.addAttribute("project", project);
-		model.addAttribute(ProjectsController.ACTIVE_NAV, ACTIVE_NAV_SETTINGS);
-		model.addAttribute("page", "remote");
-		model.addAttribute("frequencies", ProjectSyncFrequency.values());
-		projectControllerUtils.getProjectTemplateDetails(model, principal, project);
-		return "projects/settings/pages/remote";
-	}
+
 
 	/**
 	 * Request for a {@link Project} deletion page
@@ -279,79 +254,7 @@ public class ProjectSettingsController {
 		return "projects/settings/pages/metadata_templates";
 	}
 
-	/**
-	 * Update the project sync settings
-	 *
-	 * @param projectId
-	 *            the project id to update
-	 * @param frequency
-	 *            the sync frequency to set
-	 * @param forceSync
-	 *            Set the project's sync status to MARKED
-	 * @param changeUser
-	 *            update the user on a remote project to the current logged in
-	 *            user
-	 * @param principal
-	 *            The current logged in user
-	 * @param locale
-	 *            user's locale
-	 *
-	 * @return result message if successful
-	 */
-	@RequestMapping(value = "/sync", method = RequestMethod.POST)
-	@ResponseBody
-	public Map<String, String> updateProjectSyncSettings(@PathVariable Long projectId,
-			@RequestParam(required = false) ProjectSyncFrequency frequency,
-			@RequestParam(required = false, defaultValue = "false") boolean forceSync,
-			@RequestParam(required = false, defaultValue = "false") boolean changeUser, Principal principal,
-			Locale locale) {
-		Project read = projectService.read(projectId);
-		RemoteStatus remoteStatus = read.getRemoteStatus();
 
-		Map<String, Object> updates = new HashMap<>();
-
-		String message = null;
-		String error = null;
-
-		if (frequency != null) {
-			updates.put("syncFrequency", frequency);
-			message = messageSource.getMessage("project.settings.notifications.sync", new Object[] {}, locale);
-		}
-
-		if (forceSync) {
-			remoteStatus.setSyncStatus(SyncStatus.MARKED);
-			updates.put("remoteStatus", remoteStatus);
-			message = messageSource.getMessage("project.settings.notifications.sync", new Object[] {}, locale);
-		}
-
-		if (changeUser) {
-			// ensure the user can read the project
-			try {
-				projectRemoteService.read(remoteStatus.getURL());
-
-				User user = userService.getUserByUsername(principal.getName());
-				remoteStatus.setReadBy(user);
-				updates.put("remoteStatus", remoteStatus);
-
-				message = messageSource.getMessage("project.settings.notifications.sync.userchange", new Object[] {},
-						locale);
-			} catch (Exception ex) {
-				error = messageSource.getMessage("project.settings.notifications.sync.userchange.error",
-						new Object[] {}, locale);
-			}
-		}
-
-		projectService.updateProjectSettings(read, updates);
-
-		Map<String, String> response;
-		if (error == null) {
-			response = ImmutableMap.of("result", message);
-		} else {
-			response = ImmutableMap.of("error", error);
-		}
-
-		return response;
-	}
 
 	/**
 	 * Update the coverage QC setting of a {@link Project}
